feat(better_ws): utils + types top level + ws types
