+++ pragma-node/src/handlers/entries/create_entry.rs
@@ -6,10 +6,10 @@ use starknet::core::crypto::{ecdsa_verify, Signature};
 use starknet::core::types::FieldElement;
 
 use crate::config::config;
-use crate::handlers::entries::types::build_publish_message;
 use crate::handlers::entries::{CreateEntryRequest, CreateEntryResponse};
 use crate::infra::kafka;
 use crate::infra::repositories::publisher_repository;
+use crate::types::entries::build_publish_message;
 use crate::utils::JsonExtractor;
 use crate::AppState;
 
@@ -123,7 +123,7 @@ pub async fn create_entries(
 
 #[cfg(test)]
 mod tests {
-    use crate::handlers::entries::types::{BaseEntry, Entry};
+    use crate::types::entries::{BaseEntry, Entry};
 
     use super::*;
     use rstest::rstest;
