+++ pragma-node/src/handlers/entries/create_future_entry.rs
@@ -5,7 +5,8 @@ use pragma_entities::{EntryError, NewFutureEntry, PublisherError};
 use starknet::core::crypto::{ecdsa_verify, Signature};
 use starknet::core::types::FieldElement;
 
-use super::{types::build_publish_message, CreateFutureEntryRequest, CreateFutureEntryResponse};
+use super::{CreateFutureEntryRequest, CreateFutureEntryResponse};
+use crate::types::entries::build_publish_message;
 
 use crate::config::config;
 use crate::infra::kafka;
@@ -139,7 +140,7 @@ mod tests {
     use super::*;
     use rstest::rstest;
 
-    use crate::handlers::entries::types::{FutureEntry, PerpEntry};
+    use crate::types::entries::{FutureEntry, PerpEntry};
 
     #[rstest]
     fn test_build_publish_message_empty() {
