+++ pragma-node/src/handlers/entries/get_onchain/checkpoints.rs
@@ -3,10 +3,10 @@ use axum::Json;
 
 use pragma_entities::CheckpointError;
 
-use crate::handlers::entries::utils::currency_pair_to_pair_id;
 use crate::handlers::entries::{GetOnchainCheckpointsParams, GetOnchainCheckpointsResponse};
 use crate::infra::repositories::entry_repository::get_decimals;
 use crate::infra::repositories::onchain_repository::get_checkpoints;
+use crate::utils::currency_pair_to_pair_id;
 use crate::utils::PathExtractor;
 use crate::AppState;
