+++ pragma-node/src/handlers/entries/test_ws.rs
@@ -1,158 +1,20 @@
 use std::fmt::Debug;
-use std::net::IpAddr;
 use std::net::SocketAddr;
 use std::ops::ControlFlow;
 use std::sync::Arc;
-use std::time::Duration;
-use tokio::sync::mpsc::{self, Receiver};
+use tokio::sync::mpsc;
 
 use axum::extract::ws::{Message, WebSocket, WebSocketUpgrade};
 use axum::extract::{ConnectInfo, State};
 use axum::response::IntoResponse;
-use futures_util::stream::{SplitSink, SplitStream};
-use futures_util::{SinkExt, StreamExt};
+use futures_util::SinkExt;
 use serde::{Deserialize, Serialize};
 use tokio::sync::watch;
-use tokio::time::{interval, Interval};
-use uuid::Uuid;
 
+use crate::types::ws::{ChannelHandler, Subscriber};
 use crate::AppState;
 
-// ====================================================================================
-// ====================================================================================
-// ====================================================================================
-
-/// Subscriber is an actor that handles a single websocket connection.
-/// It listens to the store for updates and sends them to the client.
-pub struct Subscriber<WsState, Payload> {
-    pub id: Uuid,
-    pub ip_address: IpAddr,
-    // TODO: what is the purpose on closed?
-    pub closed: bool,
-    pub _app_state: Arc<AppState>,
-    pub _ws_state: Arc<WsState>,
-    // client receiving messages from the server
-    pub notify_receiver: Receiver<Payload>,
-    // server sending messages to the client
-    pub sender: SplitSink<WebSocket, Message>,
-    // server receiving messages from the client
-    pub receiver: SplitStream<WebSocket>,
-    pub update_interval: Interval,
-    // TODO: watches for exit but where/when is it set?
-    pub exit: watch::Receiver<bool>,
-}
-
-pub trait ChannelHandler<WsState, Payload> {
-    async fn handle_client_message(&mut self, message: Payload);
-    async fn handle_server_message(
-        &mut self,
-        subscriber: &mut Subscriber<WsState, Payload>,
-        payload: Payload,
-    );
-    async fn decode_and_handle_message_received(
-        &mut self,
-        subscriber: &mut Subscriber<WsState, Payload>,
-        msg: Message,
-    ) -> ControlFlow<(), ()>;
-    async fn periodic_interval(&mut self, subscriber: &mut Subscriber<WsState, Payload>);
-}
-
-impl<WsState, Payload> Subscriber<WsState, Payload>
-where
-    // TODO(akhercha): Remove Debug
-    WsState: Default + Debug,
-    Payload: Debug,
-{
-    pub async fn new(
-        socket: WebSocket,
-        ip_address: IpAddr,
-        app_state: Arc<AppState>,
-        notify_receiver: Receiver<Payload>,
-        exit_receiver: watch::Receiver<bool>,
-        update_interval_in_ms: u64,
-    ) -> Result<Self, String> {
-        let (sender, receiver) = socket.split();
-        let init_state = WsState::default();
-
-        let mut subscriber = Subscriber {
-            id: Uuid::new_v4(),
-            ip_address,
-            closed: false,
-            _app_state: app_state,
-            _ws_state: Arc::new(init_state),
-            notify_receiver,
-            sender,
-            receiver,
-            update_interval: interval(Duration::from_millis(update_interval_in_ms)),
-            exit: exit_receiver,
-        };
-
-        subscriber.assert_is_healthy().await?;
-
-        Ok(subscriber)
-    }
-
-    /// Perform the initial handshake with the client - ensure the channel is healthy
-    async fn assert_is_healthy(&mut self) -> Result<(), String> {
-        let ping_status = self.sender.send(Message::Ping(vec![1, 2, 3])).await;
-        if ping_status.is_err() {
-            return Err("Failed to send a ping message".to_string());
-        }
-        Ok(())
-    }
-
-    pub async fn listen<H>(&mut self, mut handler: H)
-    where
-        H: ChannelHandler<WsState, Payload>,
-    {
-        loop {
-            tokio::select! {
-                // Messages from the client
-                maybe_client_msg = self.receiver.next() => {
-                    match maybe_client_msg {
-                        Some(Ok(client_msg)) => {
-                            tracing::info!("New message from client 👇");
-                            let client_msg = handler.decode_and_handle_message_received(self, client_msg).await;
-                            if client_msg.is_break() {
-                                break;
-                            }
-                        }
-                        Some(Err(_)) => {
-                            tracing::info!("Client disconnected or error occurred. Closing the channel.");
-                            break;
-                        },
-                        None => {}
-                    }
-                },
-                // Periodic updates
-                _ = self.update_interval.tick() => {
-                    tracing::info!("🕒 [PERIODIC INTERVAL]");
-                    handler.periodic_interval(self).await;
-                },
-                // Messages from the server to the client
-                maybe_server_msg = self.notify_receiver.recv() => {
-                    if let Some(server_msg) = maybe_server_msg {
-                        tracing::info!("🥡 [SERVER MESSAGE]");
-                        handler.handle_server_message(self, server_msg).await;
-                    }
-                },
-                // Exit signal
-                _ = self.exit.changed() => {
-                    if *self.exit.borrow() {
-                        tracing::info!("⛔ [CLOSING SIGNAL]");
-                        break;
-                    }
-                },
-            }
-        }
-    }
-}
-
-// ====================================================================================
-// ====================================================================================
-// ====================================================================================
-
-struct WsTestEndpointHandler;
+struct WsTestHandler;
 
 #[derive(Debug, Clone, Deserialize, Serialize, Default)]
 pub struct WsState {
@@ -164,7 +26,7 @@ pub struct ChannelUpdateMsg {
     pub msg: String,
 }
 
-impl ChannelHandler<WsState, ChannelUpdateMsg> for WsTestEndpointHandler {
+impl ChannelHandler<WsState, ChannelUpdateMsg> for WsTestHandler {
     async fn handle_client_message(&mut self, message: ChannelUpdateMsg) {
         tracing::info!("{:?}", message);
     }
@@ -225,10 +87,6 @@ impl ChannelHandler<WsState, ChannelUpdateMsg> for WsTestEndpointHandler {
     }
 }
 
-// ====================================================================================
-// ====================================================================================
-// ====================================================================================
-
 #[utoipa::path(
     get,
     path = "/node/v1/data/test",
@@ -250,17 +108,17 @@ pub async fn test_ws(
 
 async fn create_new_subscriber(socket: WebSocket, app_state: AppState, client_addr: SocketAddr) {
     // Channel communication between the server & the subscriber
-    let (notify_sender, notify_receiver) = mpsc::channel::<ChannelUpdateMsg>(100);
+    let (_notify_sender, notify_receiver) = mpsc::channel::<ChannelUpdateMsg>(100);
     // Exit signal, allow to close the channel from the server side
-    let (exit_sender, exit_receiver) = watch::channel(false);
+    let (_exit_sender, exit_receiver) = watch::channel(false);
 
     let mut subscriber = match Subscriber::<WsState, ChannelUpdateMsg>::new(
         socket,
         client_addr.ip(),
         Arc::new(app_state),
+        1000,
         notify_receiver,
         exit_receiver,
-        1000,
     )
     .await
     {
@@ -270,10 +128,6 @@ async fn create_new_subscriber(socket: WebSocket, app_state: AppState, client_ad
             return;
         }
     };
-
-    // Save information about the current subscriber
-    let current_subscriber_id = subscriber.id;
-    let current_subscriber_ip = subscriber.ip_address;
 
     // Send a welcome message
     let _ = subscriber
@@ -285,27 +139,6 @@ async fn create_new_subscriber(socket: WebSocket, app_state: AppState, client_ad
         .await;
 
     // Main event loop for the subscriber
-    let handler = WsTestEndpointHandler;
-    tokio::spawn(async move {
-        subscriber.listen(handler).await;
-    });
-
-    // Send some messages to the client as the server using the notification chan
-    for _ in 0..50 {
-        let _ = notify_sender
-            .send(ChannelUpdateMsg {
-                msg: String::from("Hello from the server"),
-            })
-            .await;
-        // wait 5s
-        tokio::time::sleep(Duration::from_secs(10)).await;
-    }
-
-    // close the channel when talk is over
-    tracing::info!(
-        "End of discussions. Closing channel for [{}] {}",
-        current_subscriber_ip,
-        current_subscriber_id
-    );
-    let _ = exit_sender.send(true);
+    let handler = WsTestHandler;
+    subscriber.listen(handler).await;
 }
