+++ pragma-node/src/types/entries.rs
