+++ pragma-node/src/types/mod.rs
@@ -1,3 +1,6 @@
+pub mod entries;
+pub mod ws;
+
 /// The number of seconds since the Unix epoch (00:00:00 UTC on 1 Jan 1970). The timestamp is
 /// always positive, but represented as a signed integer because that's the standard on Unix
 /// systems and allows easy subtraction to compute durations.
